+++ internal/controller/operator/factory/vlcluster/vmauth_lb.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"sort"
+	"strings"
 
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -92,26 +93,25 @@ func buildVMauthLBSecret(cr *vmv1.VLCluster) *corev1.Secret {
 	selectURL := fmt.Sprintf("%s://%s.%s:%s",
 		selectProto, cr.PrefixedInternalName(vmv1beta1.ClusterComponentSelect), targetHostSuffix, selectPort)
 
-	lbScrt := &corev1.Secret{
-		ObjectMeta: buildLBConfigSecretMeta(cr),
-		// TODO: add backend auth
-		StringData: map[string]string{"config.yaml": fmt.Sprintf(`
+	lbConfig := fmt.Sprintf(`
 unauthorized_user:
   url_map:
   - src_paths:
     - "/insert/.*"
     - "/internal/insert"
-    url_prefix: "%s"
+    url_prefix: %q
     discover_backend_ips: true
   - src_paths:
     - ".*"
-    url_prefix: "%s"
+    url_prefix: %q
     discover_backend_ips: true
-      `, insertURL,
-			selectURL,
-		)},
+      `, insertURL, selectURL)
+
+	return &corev1.Secret{
+		ObjectMeta: buildLBConfigSecretMeta(cr),
+		// TODO: add backend auth
+		Data: map[string][]byte{"config.yaml": []byte(strings.TrimSpace(lbConfig))},
 	}
-	return lbScrt
 }
 
 func buildVMauthLBDeployment(cr *vmv1.VLCluster) (*appsv1.Deployment, error) {